Split html_generator.py into per-report generator modules
html_generator.py had grown to 2,034 lines with four report generators
and their shared helpers in a single class. Split it as a pure move
(method bodies unchanged):

- report_base.py: ReportGeneratorBase with shared state and the common
  helpers (_analyze_multiple_environments, _generate_environment_section,
  _get_platform_info, _get_common_css)
- overview_report.py, results_matrix_report.py, raw_data_report.py,
  data_index_report.py: one generator class per report type
- html_generator.py: HTMLGenerator kept as a ~130-line facade delegating
  to the four generators; public API (generate_all_reports,
  generate_overview, ...) is unchanged

Verified by golden-file comparison: reports generated from the same
seeded database before and after the split are byte-identical except
for generation timestamps. Test suite and 'main.py --validate' pass.
