+++ scripts/reporting/data_index_report.py
@@ -1,0 +1,270 @@
+"""
+Data export index report generator (extracted from html_generator.py).
+"""
+
+import sys
+from datetime import datetime, timezone
+from pathlib import Path
+
+# Add project root to path for imports
+project_root = Path(__file__).parent.parent.parent
+sys.path.insert(0, str(project_root))
+
+from scripts.reporting.report_base import ReportGeneratorBase
+
+
+class DataIndexReportGenerator(ReportGeneratorBase):
+    """Generates the data export index page (data/index.html)"""
+
+    def generate_data_index(self) -> str:
+        """Generate index.html for data directory with links to data files"""
+
+        self.logger.info("Generating data index page...")
+
+        html_content = f"""<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Optimization Solver Benchmark - Data Exports</title>
+    <style>
+        * {{
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
+        }}
+
+        body {{
+            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif;
+            line-height: 1.6;
+            color: #333;
+            background-color: #f8f9fa;
+        }}
+
+        header {{
+            background: linear-gradient(135deg, #2c3e50 0%, #3498db 100%);
+            color: white;
+            text-align: center;
+            padding: 2rem 1rem;
+            margin-bottom: 2rem;
+        }}
+
+        header h1 {{
+            font-size: 2.5rem;
+            margin-bottom: 0.5rem;
+        }}
+
+        header p {{
+            font-size: 1.1rem;
+            opacity: 0.9;
+        }}
+
+        nav {{
+            background: white;
+            padding: 1rem;
+            text-align: center;
+            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+            margin-bottom: 2rem;
+        }}
+
+        nav a {{
+            color: #2c3e50;
+            text-decoration: none;
+            margin: 0 1rem;
+            padding: 0.5rem 1rem;
+            border-radius: 4px;
+            transition: background-color 0.2s;
+        }}
+
+        nav a:hover {{
+            background-color: #ecf0f1;
+        }}
+
+        nav a.active {{
+            background-color: #3498db;
+            color: white;
+        }}
+
+        main {{
+            max-width: 1000px;
+            margin: 0 auto;
+            padding: 0 1rem;
+        }}
+
+        .section {{
+            background: white;
+            margin: 2rem 0;
+            border-radius: 8px;
+            box-shadow: 0 2px 10px rgba(0,0,0,0.1);
+        }}
+
+        .section h2 {{
+            background: linear-gradient(135deg, #34495e 0%, #2c3e50 100%);
+            color: white;
+            padding: 1rem 1.5rem;
+            margin: 0;
+            border-radius: 8px 8px 0 0;
+            font-size: 1.5rem;
+        }}
+
+        .section-content {{
+            padding: 1.5rem;
+        }}
+
+        .data-file {{
+            background: #f8f9fa;
+            border: 1px solid #e9ecef;
+            border-radius: 6px;
+            padding: 1.5rem;
+            margin: 1rem 0;
+            transition: box-shadow 0.2s;
+        }}
+
+        .data-file:hover {{
+            box-shadow: 0 4px 12px rgba(0,0,0,0.1);
+        }}
+
+        .data-file h3 {{
+            color: #2c3e50;
+            margin-bottom: 0.5rem;
+        }}
+
+        .data-file p {{
+            color: #7f8c8d;
+            margin-bottom: 1rem;
+        }}
+
+        .download-btn {{
+            display: inline-block;
+            background: #3498db;
+            color: white;
+            padding: 0.75rem 1.5rem;
+            text-decoration: none;
+            border-radius: 4px;
+            font-weight: 600;
+            transition: background-color 0.2s;
+        }}
+
+        .download-btn:hover {{
+            background: #2980b9;
+        }}
+
+        .file-size {{
+            font-size: 0.9rem;
+            color: #95a5a6;
+            margin-left: 1rem;
+        }}
+
+        footer {{
+            text-align: center;
+            padding: 2rem;
+            color: #7f8c8d;
+            border-top: 1px solid #ecf0f1;
+            margin-top: 3rem;
+        }}
+
+        footer a {{
+            color: #3498db;
+            text-decoration: none;
+            margin: 0 1rem;
+        }}
+
+        footer a:hover {{
+            text-decoration: underline;
+        }}
+    </style>
+</head>
+<body>
+    <header>
+        <h1>🔬 Optimization Solver Benchmark</h1>
+        <p>Data Exports - Download Benchmark Results</p>
+        <p><small>Generated: {datetime.now(timezone.utc).strftime("%Y-%m-%d %H:%M:%S UTC")}</small></p>
+    </header>
+
+    <nav>
+        <a href="../index.html">Overview</a>
+        <a href="../results_matrix.html">Results Matrix</a>
+        <a href="../raw_data.html">Raw Data</a>
+        <a href="index.html" class="active">Data Exports</a>
+    </nav>
+
+    <main>
+        <div class="section">
+            <h2>📁 Available Data Files</h2>
+            <div class="section-content">
+                <p style="margin-bottom: 2rem; color: #7f8c8d;">
+                    Download complete benchmark results in various formats for further analysis and research.
+                </p>
+
+                <div class="data-file">
+                    <h3>📊 Latest Results (JSON)</h3>
+                    <p>Most recent benchmark results with solver information, problem details, and performance metrics. Best for analyzing current solver performance.</p>
+                    <a href="benchmark_results_latest.json" class="download-btn">📥 Download Latest JSON</a>
+                    <span class="file-size">Latest run only</span>
+                </div>
+
+                <div class="data-file">
+                    <h3>📈 Latest Results (CSV)</h3>
+                    <p>Most recent benchmark results in CSV format for spreadsheet analysis. Contains the latest solver-problem combinations.</p>
+                    <a href="benchmark_results_latest.csv" class="download-btn">📥 Download Latest CSV</a>
+                    <span class="file-size">Latest run only</span>
+                </div>
+
+                <div class="data-file">
+                    <h3>🗄️ All Results (JSON)</h3>
+                    <p>Complete historical benchmark results including all runs. Sorted by ID for database restoration. Ideal for comprehensive analysis and backup.</p>
+                    <a href="benchmark_results_all.json" class="download-btn">📥 Download All JSON</a>
+                    <span class="file-size">Full database export</span>
+                </div>
+
+                <div class="data-file">
+                    <h3>📑 All Results (CSV)</h3>
+                    <p>Complete historical benchmark results in CSV format. Sorted by ID for database restoration. Perfect for time-series analysis and research.</p>
+                    <a href="benchmark_results_all.csv" class="download-btn">📥 Download All CSV</a>
+                    <span class="file-size">Full database export</span>
+                </div>
+
+            </div>
+        </div>
+
+        <div class="section">
+            <h2>📖 Data Format Documentation</h2>
+            <div class="section-content">
+                <h3>JSON Structure</h3>
+                <ul style="margin: 1rem 0; color: #34495e; padding-left: 1.5rem;">
+                    <li><strong>benchmark_results_latest.json</strong>: Latest benchmark run with metadata, summary statistics, and results array</li>
+                    <li><strong>benchmark_results_all.json</strong>: Complete database export sorted by ID for restoration purposes</li>
+                    <li>Each contains: metadata, summary statistics, solver comparison, and detailed results</li>
+                </ul>
+
+                <h3>CSV Format</h3>
+                <ul style="margin: 1rem 0; color: #34495e; padding-left: 1.5rem;">
+                    <li><strong>benchmark_results_latest.csv</strong>: Latest run in tabular format, sorted by problem and solver</li>
+                    <li><strong>benchmark_results_all.csv</strong>: All historical data sorted by ID (ascending) for database restoration</li>
+                    <li>Headers include: id, solver_name, solver_version, problem_library, problem_name, problem_type, solve_time, status, etc.</li>
+                </ul>
+            </div>
+        </div>
+    </main>
+
+    <footer>
+        <p>
+            <a href="../index.html">Overview</a> |
+            <a href="../results_matrix.html">Results Matrix</a> |
+            <a href="../raw_data.html">Raw Data</a>
+        </p>
+        <p><small>Generated by Optimization Solver Benchmark System</small></p>
+    </footer>
+</body>
+</html>"""
+
+        # Save to data directory
+        data_dir = self.output_dir / "data"
+        data_dir.mkdir(parents=True, exist_ok=True)
+        output_file = data_dir / "index.html"
+
+        with open(output_file, "w") as f:
+            f.write(html_content)
+
+        self.logger.info(f"Data index page saved to {output_file}")
+        return html_content
